requirements.txt
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
astor==0.8.1
autograd==1.7.0
autograd-gamma==0.5.0
backports.zoneinfo==0.2.1
certifi==2025.1.31
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
distro==1.9.0
Django==4.2.19
django-allauth==65.4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
exceptiongroup==1.2.2
fonttools==4.55.8
formulaic==1.1.1
graphlib_backport==1.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_resources==6.4.5
inflection==0.5.1
interface-meta==1.3.0
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.7
lifelines==0.27.8
matplotlib==3.7.5
numpy==1.24.4
openai==1.63.0
packaging==24.2
pandas==2.0.3
patsy==1.0.1
pillow==10.4.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
statsmodels==0.14.1
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
wrapt==1.17.2
zipp==3.20.2
